Refactor site LoginPage, LoginPageXPath, Login Test
